Adding hash cli
Adding dep

Adding dumpca command for non-windows only

Adding ability to specify a host:port to connect to, in addition
to file based cert bundle.
