+++ system_specific_cmds_windows.go
@@ -1,0 +1,5 @@
+package main
+
+import "github.com/mitchellh/cli"
+
+func system_specific_cmds(cmds map[string]cli.CommandFactory) {}
